card repository and services added In addition to dtos and payment with wallet added to CustomerServiceImpl

+++ src/main/java/finalproject/finalproject/Entity/Wallet.java
@@ -13,6 +13,6 @@ import lombok.experimental.SuperBuilder;
 @Getter
 @Setter
 public class Wallet extends BaseEntity<Integer> {
-    private int creditOfWallet;
+    private double creditOfWallet;
 
 }
